In the controller test directory, create a new class: BattleControllerTest. Make it extend the fancy ApiTestCase we've been working on.
Start with public function testPOSTCreateBattle().

Go steal some code from ProgrammerControllerTest(). Copy the setup method that creates a user so that we can send the Authorization header.

Create a Project & Programmer

For this endpoint, there are only two pieces of information we need to send: which programmer and which project will battle. So before we start, we need to create these. Add{dollar_sign}this->createProject() and give it a name: my_project - that doesn't matter. If you open this method, this method simply creates a new Project and flushes it to the database.

Next, create the programmer: {dollar_sign}this->createProgrammer(). This takes an array of information about that programmer. Hmm, let's call him Fred. Pass weaverryan as the second argument: that will be the user who owns Fred.

Eventually, we'll need to restrict this endpoint so that we can only start battles with our programmers.