Sending the Related Fields
Next, send a POST request! In ProgrammerControllerTest, it was easy: we sent 3 scalar fields.
But now, it's a little different: we want to send data that identifies the related programmer and project resources. Should we send the id? Or the programmer's nickname?

Well, like normal with API's... it doesn't matter. But sending the id's makes sense. Create a new {dollar_sign}data array with two fields: project set to {dollar_sign}project->getId() and programmer set to {dollar_sign}programmer->getId().

I'm calling the keys programmer and project for obvious reasons: but they could be anything. We are in charge of naming the fields whatever we want. Just be sane and consistent: please don't call the fields bob and larry - everyone will hate you.

Finally, make the request: {dollar_sign}response = {dollar_sign}this->client->post(). For programmers, the URL is /api/programmers. Stay consistent with /api/battles. Pass an array as the second argument with a body key set to json_encode({dollar_sign}data) and a headers key set to {dollar_sign}this->getAuthorizedHeaders('weaverryan').

That will send a valid JSON web token for the weaverryan user - we created that in the previous course.