Asserts!
Whew, okay. That's it. So even though Battle is dependent on two other resources, it works pretty much the same. Add some basic asserts: {dollar_sign}this->assertEquals() that 201 will be the response status code.

In Battle, one of the fields that should be returned is didProgrammerWin. Make sure that exists with {dollar_sign}this->asserter()->assertResponsePropertyExists() and look for didProgrammerWin.

I'll also add a todo to check for the Location header later. Remember, when you create a resource, you're supposed to return a Location header to the URL where the client can view the new resource. We don't have a GET endpoint for a battle yet, so we'll skip this.

The hard work is done: we've designed the new endpoint. Let's bring it to life!