BattleManager Complicates Things...
Well, not so simple in this case. What? I know, I like to make things as difficult as possible.

To create battles on the frontend, our controller uses a service class called BattleManager. It's kind of nice: it has a battle() function:

 55 lines  src/AppBundle/Battle/BattleManager.php
... lines 1 - 9
class BattleManager {
... lines 12 - 18
  /**
   * Creates and wages an epic battle
   *
   * @param Programmer {dollar_sign}programmer
   * @param Project {dollar_sign}project
   * @return Battle
   */
  public function battle(Programmer {dollar_sign}programmer, Project {dollar_sign}project){
... lines 28 - 53
  }
}
We pass it a Programmer and Project and it takes care of all of the logic for creating a Battle, figuring out who won, and saving it to the database. We even gave Battle a __construct() function with two required arguments:

 107 lines  src/AppBundle/Entity/Battle.php
... lines 1 - 10
class Battle {
... lines 13 - 46
  /**
   * Battle constructor.
   * @param {dollar_sign}programmer
   * @param {dollar_sign}project
   */
  public function __construct(Programmer {dollar_sign}programmer, Project {dollar_sign}project){
    {dollar_sign}this->programmer = {dollar_sign}programmer;
    {dollar_sign}this->project = {dollar_sign}project;
    {dollar_sign}this->foughtAt = new \DateTime();
  }
... lines 58 - 105
}
This is a really nice setup, so I don't want to change it. But, it doesn't work well with the form system: it prefers to instantiate the object and use setter functions.

 Tip
Actually, it is possible to use the form system with the Battle entity by taking advantage of data mappers(https://webmozart.io/blog/2015/09/09/value-objects-in-symfony-forms/).

But that's ok! In fact, I like this complication: it shows off a very nice workaround. Just create a new model class for the form. In fact, I recommend this whenever you have a form that stops looking like or working nicely with your entity class.