Adding the BattleModel
In the Form directory, create a Model directory to keep things organized. Inside, add a new class called BattleModel.
Give it the two properties we're expecting as API input:  and {dollar_sign}programmer. Hit command+N - or go to the 'Code'->'Generate' menu - and generate the getter and setter methods for both properties.

To be extra safe and make your code more hipster, type-hint setProgrammer() with the Programmer class and setProject() with Project. The form system will love this class.

Designing the Form

In the Form directory, create a new class for the form: BattleType. Make this extend the normal AbstractType and then hit command+N - or 'Code'->'Generate' - and go to 'Override Methods'. Select the two we need: buildForm and configureOptions.

Take out the parent calls - the parent methods are empty.

EntityType to the Rescue!

Okay, let's think about this. The API client will send programmer and project fields and both will be ids. Ultimately, we want to turn those into the entity objects corresponding to those ids before setting the data on the BattleModel object.

Well, this is exactly what the Entity type does. Use {dollar_sign}builder->add() with project set to EntityType::class. To tell it what entity to use, add a class option set to AppBundle\Entity\Project.
Do the same for programmer and set its class to AppBundle\Entity\Programmer.

In a web form, the entity type renders as a drop-down of projects or programmers. But it's perfect for an API: it transforms the project id into a Project object by querying for it. That's exactly what we want.

In configureOptions(), add {dollar_sign}resolver->setDefaults() and pass it two things: first the data_class set to BattleModel::class.

Make sure PhpStorm adds the use statement for that class. Then, set csrf_protection to false because we can't use normal CSRF protection in an API.
Form, ready! Now let's hit the controller.