Head to the controller. We've got this form flow mastered! Step 1: create a new BattleModel object: {dollar_sign}battleModel = new BattleModel(). Step 2: create the form: {dollar_sign}form = {dollar_sign}this->createForm() with BattleType::class. For the second argument to createForm: pass it {dollar_sign}battleModel.
Step 3: Use {dollar_sign}this->processForm(). Remember, this is a method we added in BaseController:
87 lines  src/AppBundle/Controller/BaseController.php
... lines 1 - 19
abstract class BaseController extends Controller {
... lines 22 - 142
  protected function processForm(Request {dollar_sign}request, FormInterface {dollar_sign}form){
  {dollar_sign}data = json_decode({dollar_sign}request->getContent(), true);
  if ({dollar_sign}data === null) {
    {dollar_sign}apiProblem = new ApiProblem(400, ApiProblem::TYPE_INVALID_REQUEST_BODY_FORMAT);
      throw new ApiProblemException({dollar_sign}apiProblem);
    }
    {dollar_sign}clearMissing = {dollar_sign}request->getMethod() != 'PATCH';
    {dollar_sign}form->submit({dollar_sign}data, {dollar_sign}clearMissing);
  }
... lines 155 - 185
}
it decodes the Request body and submits it into the form, which is what we do on every endpoint that processes data.

Type-hint the Request argument for the controller and pass this to processForm().

If the form is not valid, we need to send back errors. Use another method from earlier: {dollar_sign}this->throwApiProblemValidationException() and pass it the {dollar_sign}form object.
This will grab the validation errors off and create that response.