At this point, we have a BattleModel object that's populated with the Programmer and Project objects sent in the request. To create the battle, we need to use the BattleManager. Do that with {dollar_sign}this->getBattleManager() - that's just a shortcut to get the service - ->battle() and pass it {dollar_sign}battleModel->getProgrammer() and {dollar_sign}battleModel->getProject().
Put a little {dollar_sign}battle = in the beginning of all of this to get the new Battle object. Perfect!

Now that the battle has been heroically fought, let's send back the gory details. Use return {dollar_sign}this->createApiResponse() and pass it {dollar_sign}battle and the 201 status code.

We aren't setting a Location header yet, so let's at least add a todo for that.

We are done! Controller, model and form: these are the only pieces we need to create a robust endpoint. Try the test:

./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
We are in prime battling shape.

Now, let's complicate things and learn how to really take control of every field in our endpoint. And, learn more about relations.