The test passes, but let's see what the response looks like. Add {dollar_sign}this->debugResponse(). and re-run the test:
./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Check it out! It has the fields we expect, but it's also embedding the entire programmer and project resources. That's what the serializer does when a property is an object. This might be cool with you, or maybe not. For me, this looks like overkill. Instead of having the Programmer and Project data right here, it's probably enough to just have the programmer's nickname and the project's id.

But hold on: I want to mention something really important. Whenever you need to make a decision about how your API should work, the right decision should always depend on who you're making the API for. If you're building your API for an iPhone app, will having these extra fields be helpful? Or, if you're API is for a JavaScript frontend like ReactJS, then build your API to make React happy.