Adding Fake Properties
Let's go to the next level. Now, I do want to have a programmer, but set to the username instead of the whole object. And I also do want a project field, set to its id.

Update the test to look for these. Use {dollar_sign}this->asserter()->assertResponsePropertyEquals() and pass it {dollar_sign}response. Look for a project field that's set to {dollar_sign}project->getId(). Copy that line and do the same thing for programmer: it should equal Fred.

We could also have this return the id - it's up to you and what's best for your client.

But, how can we bring this to life? We're in a weird spot, because these fields do exist on Battle, but they have the wrong values. How can we do something custom?