By using something called a virtual property. First, create a new public function called getProgrammerNickname(). It should return {dollar_sign}this->programmer->getNickname().
VirtualProperty

Simple. But that will not be used by the serializer yet. To make that happen, add @Serializer\VirtualProperty above the method. As soon as you do this, it will be exposed in your API. But it will be called programmerNickname the serializer generates the field name by taking the method name and removing get.

SerializedName

Since we want this to be called programmer add another annotation: @Serializer\SerializedName() and pass it programmer.

Now we have a programmer field set to the return value of this method.

Do the same thing for project: public function getProjectId(). This will return {dollar_sign}this->project->getId(). Above this, add the @Serializer\VirtualProperty to activate the new field and @Serializer\SerializedName('project') to control its name.

Head to the terminal and try the test:

./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
We've got it! This trick is a wonderful way to take control of exactly how you want your representation to look.