Remember, we want to design our API to work well with whomever is using it: whether that's a third-party API client, a JavaScript front end, or another PHP app that's talking to us. That's why we just changed how our Battle output looks.
But you might also want to control how the input looks: what the client needs to send to your API. For example, before, to create a new battle, you send a project field and a programmer field: each set to their ID.

But what if we wanted to call these fields projectId and programmerId?

After all, those are IDs that are being sent. If we change this in the test, everything will explode. Prove it by running things:

./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Yep, a big validation error: the form should not contain extra fields: these two new fields are not in the form we built.