Using property_path
Here is the very simple, elegant, amazing solution. In the form, you do need to update your fields to projectId and programmerId so they match what the client is sending. But then, add a property_path option to projectId set to project. Do the same thing to the programmerId field: 'property_path' => 'programmer'.

That's the key! The form now expects the client to send projectId and programmerId. But when it sets the final data on BattleModel, it will call setProject() and setProgrammer().

This is a little known way to have a field name that's different than the property name on your class. Bring on the test!

./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Awesome! Another useful option I want you to know about is called mapped. You can use this to allow an extra field in your input, without needing to add a corresponding property to your class.