It doesn't make any sense to create a battle without a Programmer or a Project. But guess what - you can! Or at least, you kind of can: we don't have validation to prevent that yet!
The validation system we created in earlier courses is air-tight: as long as we add the constraint annotations, it just works. So normally, I might not write a test for failing validation. But I will now... because we're going to add a twist.

Testing for Validation

Add a new public function testPOSTBattleValidationErrors(). Copy the first bits from the previous function that create the data and make the request.

But, don't actually create a project! Instead, send null for the projectId. Since starting a battle against nothing is nonsense, assert that 400 is the response status code. This follows the pattern we did before in ProgrammerControllerTest.

And actually, that test shows off the validation errors response format: there should be an errors key with field names for the errors below that. Each field could technically have multiple errors, so that's an array.

Check for the error in our code with {dollar_sign}this->asserter()->assertResponsePropertyExists(): the field should be errors.projectId.

Next, check for the exact message: assertResponsePropertyEquals() with errors.projectId[0] - so the first and only error - set to This value should not be blank.

Why that message? That's the default message for Symfony's NotBlank constraint.

Before we code this up, copy the method name and run the test:

./vendor/bin/phpunit --filter testPOSTBattleValidationErrors src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
It explodes with a 500 error! This is what happens when you're lazy and forget to add validation: the BattleManager panics because there is no Project. We do not want 500 errors, they are not hipster.