Adding Basic Validation
We know how to fix this! Go to BattleModel. Remember, this is the class that's bound to the form: so the annotations should go here. First, add the use statement. Type use NotBlank, let it auto-complete, delete the last part and add the normal as Assert.

That's my shortcut to get the use statement.

Now, above project, add @Assert\NotBlank(). Do the same above programmer: @Assert\NotBlank().

Done! Now run the test:

./vendor/bin/phpunit --filter testPostBattleValidationErrors src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
We're awesome! Or are we... there's a deeper problem! What prevents an API client from starting a battle with a Programmer that they do not own? Right now - nothing, besides karma and trusting that humankind will do the right thing. Unfortunately, that doesn't usually pass a security audit. Let's be heros and fix this security hole!