Perfect! Now, let's go add that link. First, open up the Programmer entity. We added the self link earlier via a cool annotation system we created.
In Battle, add something similar: @Link - let that auto-complete for the use statement. Set the name - or rel - of the link to programmer. This is the significance of the link: it could be anything, as long as you consistently use programmer when linking to a programmer.

For the route, use api_programmers_show: the route name to a single programmer. Finally, add params: the wildcard parameters that need to be passed to the route. This route has a nickname wildcard. Set it to an expression: object.getProgrammerNickname(). That's the method we created down below earlier.

And that's all we need. Copy the method name again - testPostCreateBattle() - and run the test:

./vendor/bin/phpunit --filter testPostCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
And it works.

Now, let me show you an awesome library that makes adding links even easier. In fact, I stole the @Link annotation idea from it.