In the test, the Programmer is owned by weaverryan and then we authenticate as weaverryan. So, we're starting a battle using a Programmer that we own. Time to mess that up. Create a new user called someone_else. There still is a user called weaverryan.
But now, change the programmer to be owned by this sketchy someone_else character.
With this setup, weaverryan will be starting a battle with someone_else's programmer. This should cause a validation error: this is an invalid programmerId to pass.

Form Field Sanity Validation

But how do we do that? Is there some annotation we can use for this? Nope! This validation logic will live in the form. 'What!?' you say - 'Validation always goes in the class! '. Not true! Every field type has a little bit of built-in validation logic. For example, the NumberType will fail if a mischievous - or confused - user types in a word. And the EntityType will fail if someone passes an id that's not found in the database. I call this sanity validation: the form fields at least make sure that a sane value is passed to your object.

If we could restrict the valid programmer id's to only those owned by our user, we'd be in business.

But first, add the test: assertResponsePropertyEquals() that errors.programmerId[0] should equal some dummy message.
Run the test to see the failure:

./vendor/bin/phpunit --filter testPOSTBattleValidationErrors src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Yep: there's no error for programmerId yet.

Let's fix that. Right now, the client can pass any valid programmer id, and the EntityType happily accepts it. To shrink that to a smaller list, we'll pass it a custom query to use.