Passing the User to the Form To do that, the form needs to know who is authenticated. In BattleController, guarantee that first: add {dollar_sign}this->denyAccessUnlessGranted('ROLE_USER'). To pass the user to the form, add a third argument to createForm(), which is a little-known options array. Invent a new option: user set to {dollar_sign}this->getUser().
This isn't a core Symfony thing: we're creating a new option.

To allow this, open BattleType and find configureOptions. Here, you need to say that user is an allowed option. One way is via {dollar_sign}resolver->setRequired('user'). This means that whoever uses this form is allowed to, and in fact must, pass a user option.

With that, you can access the user object in buildForm(): {dollar_sign}user = {dollar_sign}options['user'].

None of this is unique to API's: we're just giving our form more power!