Passing the query_builder Option Let's filter the programmer query: add a query_builder option set to an anonymous function with ProgrammerRepository as the only argument. Add a use for  so we can access it.
We could write the query right here, but you guys know I don't like that: keep your queries in the repository! Call a new method createQueryBuilderForUser() and pass it {dollar_sign}user.

Copy that method name and shortcut-your way to that class by holding command and clicking ProgrammerRepository. Add public function createQueryBuilderForUser() with the User {dollar_sign}user argument. Inside, return {dollar_sign}this->createQueryBuilder() and alias the class to programmer. Then, just andWhere('programmer.user = :user') with ->setParameter('user', {dollar_sign}user).
Done! The controller passes the User to the form, and the form calls the repository to create the custom query builder. Now, if someone passes a programmer id that we do not own, the EntityType will automatically cause a validation error. Security is built-in.

Head back to the terminal to try it!

./vendor/bin/phpunit --filter testPOSTBattleValidationErrors src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Awesome! Well, it failed - but look! It's just because we don't have the real message yet: it returned This value is not valid.. That's the standard message if any field fails the 'sanity' validation.

 Tip
You can customize this message via the invalid_message form field option.