Adding the HATEOAS Annotation Links
Go back to the library itself, and scroll down to the first coding example. This uses an annotation system that looks pretty similar to ours. Copy the use statement on top, open Battle and paste that.
Next, change the annotation to @Hateoas\Relation.

Keep programmer: that will still be the link's rel. But add href=@Hateoas\Route and pass that the name of the route: api_programmers_show. Update params to parameters, and inside, set nickname equal, and wrap the expression in expr().

That translates our annotation format to the one used by the bundle. And the result is almost the same. Open BattleControllerTest and copy the first method name.

we have a test for a link near the bottom of this:

 76 lines  tests/AppBundle/Controller/Api/BattleControllerTest.php
... lines 1 - 6
class BattleControllerTest extends ApiTestCase
{
... lines 9 - 15
  public function testPOSTCreateBattle(){
... lines 18 - 40
    {dollar_sign}this->asserter()->assertResponsePropertyEquals(
      {dollar_sign}response,
      '_links.programmer',
      {dollar_sign}this->adjustUri('/api/programmers/Fred')
    );
... lines 46 - 48
  }
... lines 50 - 74
}

Change over to the terminal and, as long as Composer is done, run:

vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Check it out! It fails - but barely. This library still puts links under an _links key, but instead of listing the URLs directly, it wraps each inside an object with an href key. That's causes the failure.