Ok, fair enough. Let's fix that by updating the test to look for _links.programmer.href.
Run the test again:

vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
And now we're green.

Holy Toledo Batman: This is HAL JSON!

But guess what? It's no accident that this library used this exact format: with an _links key and an href below that. This is a semi-official standard format called HAL JSON.