Advertising that you're using Hal
So now that we are using Hal, we should advertise it! In fact, that's what this application/hal+json means in their documentation: it's a custom Content-Type. It means that the format is JSON, but there's some extra rules called Hal. If a client sees this, they can Google for it.

In ProgrammerControllerTest, assert that application/hal+json is equal to {dollar_sign}response->getHeader('Content-Type')[0].

Guzzle returns an array for each header - there's a reason for that, but yea, I know it looks ugly.

To actually advertise that our API returns HAL, open BaseController and search for createApiResponse() - the method we're calling at the bottom of every controller. Change the header to be application/hal+json.

Nice! Copy the test name and re-run the test:

./vendor/bin/phpunit --filter testPOSTProgrammerWorks src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
Congratulations! Your API is no longer an island: welcome to the club.