Adding an Embedded Relation
To do that, after href, add an embedded='expr()' with object.getProgrammer().
Let's see what this looks like! Open BattleControllerTest and right at the bottom, add our handy {dollar_sign}this->debugResponse({dollar_sign}response).

Perfect! Copy that method name and run it:

./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest.php
Oh, cool: we still have the relation in _links, but now we also have an entire programmer resource in _embedded. So when you setup @Hateoas\Relation annotations you can choose whether you want this to be a link or an embedded object.