Custom Serializer
In AppBundle, in the Serializer directory, create a new class called CustomHATEOASJsonSerializer

Make it extend a class called JsonHalSerializer: this is the current class responsible for adding links in the HAL format. In fact: open up the class.

It has two method. serializeLinks() is responsible for reading the Relation annotations and adding them to the JSON with _links. serializeEmbeddeds() adds any embedded relations under the _embedded key.

For now, let's focus on changing how the links render only. Go to the 'Code'->'Generate' menu - command+N on a Mac - and hit 'Override Methods'. Override serializeLinks().

Re-open the parent method and then the interface: I want to copy all that good PHPDoc so we get auto-complete. Paste it above our method and auto-complete the Link to get its use statement.

Alright: this should be easy.

Create a {dollar_sign}serializedLinks array and foreach over the {dollar_sign}links variable. Each of these is a Link object, and contains the configuration for one annotation. Now, just create the format we want: {dollar_sign}serializedLinks[], with {dollar_sign}link->getRel(). Instead of setting this to an array with an href key, simply set it to {dollar_sign}link->getHref().
Perfect! Finally, at the bottom, we need to add the _links property. Do that with: {dollar_sign}visitor->addData('_links', {dollar_sign}serializedLinks).
With any luck, that should give us a simpler format without that href.