Changing our Tests Back
But don't run the tests quite yet: we know some things will be broken. In BattleControllerTest, take off the href we just added: it should be _links.programmer. And in ProgrammerControllerTest, under testGETProgrammer, do the same.

Phew! That's a lot of changes, so let's re-run the entire test suite:

./vendor/bin/phpunit
Hey, it passes! I must've left a debugResponse() in there somewhere: but that's nothing to worry about - we're green!