Let me show you something else I don't really like. In BattleControllerTest, we're checking for the embedded programmer. Right now it's hidden under this _embedded key.
Hal does this so that a client knows which data is for the Battle, and which data is for the embedded programmer. But what if it would be more convenient for our client if the data was not under an _embedded key? What if they want the data on the root of the object like it was before?

Well, that's fine! Just stop using the embedded functionality from the bundle. Delete the assert that looks for the string and instead assert that the programmer.nickname is equal to Fred.

In other words, I want to change the root programmer key from a string to the whole object. And we'll eliminate the _embedded key entirely.

In Battle.php, remove the embedded key from the annotation.

OK, _embedded is gone! Next, on the programmer property, add @Expose.
The serializer will serialize that whole object. We originally didn't expose that property because we added this cool @VirtualProperty above the getProgrammerNickname() method.

In BattleControllerTest, let's see if this is working. First dump the response. Copy the method name, and give this guy a try:
./vendor/bin/phpunit --filter testPOSTCreateBattle src/AppBundle/Tests/Controller/Api/BattleControllerTest

It passes! And the response looks exactly how we want: no more _embedded key.