Adding Real Pagination And really: if we're going to all of this trouble to use the PaginatedCollection, shouldn't we go one extra half-step and actually add pagination? After all, it'll make this endpoint even more consistent by having those pagination links.
Change the {dollar_sign}collection = line to {dollar_sign}this->get('pagination_factory')->createCollection().

This needs a few arguments. The first is a query builder. So instead of making this full query for battles, we need to just return the query builder. Rename this to a new method called - createQueryBuilderForProgrammer() and pass it the{dollar_sign}programmer object.

I'll hold command and I'll click Battle to jump into BattleRepository. Add that method: public function createQueryBuilderForProgrammer() with a Programmer {dollar_sign}programmer argument.

Fortunately, the query is easy: return {dollar_sign}this->createQueryBuilder('battle'), then ->andWhere('battle.programmer = :programmer') with setParameter('programmer', {dollar_sign}programmer)

Perfect! Back in ProgrammerController, rename the variable to {dollar_sign}battlesQb and pass it to createCollection(). The second argument is the request object. You guys know what to do: type-hint a new argument with Request and pass that in. The third argument is the name of the route the pagination links should point to. That's this route: api_programmers_battles_list. Finally, the last argument is any route parameters that need to be passed to the route. This route has a nickname, so pass nickname => {dollar_sign}programmer->getNickname()

Done. We basically changed one line to create a real paginated collection. And now, we celebrate. Run the test:

./vendor/bin/phpunit --filter testFollowProgrammerBattlesLink src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
That is real pagination pretty much out of the box. Yea, this only has three results and only one page: but if this programmer keeps having battles, we're covered.

We've really perfected a lot of traditional REST endpoints. Now, let's talk about what happens when endpoints get weird...